refactor(analysis): migrate analyze module Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
