+++ src/pyrovelocity/analysis/__init__.py
@@ -1,0 +1,6 @@
+import pyrovelocity.analysis.analyze
+
+
+__all__ = [
+    "analyze",
+]
